Match_AI: widen confidence to 3 decimal places (millipercent wire field)
conf_pct was u8 (0-100), rounding low-confidence predictions to whole
percent and often clipping to 100% near the calibrated tau threshold.
Wire field is now u16 millipercent (0-100000); display shows %.3f%%.

+++ src/modules/ais/ais_meta.hpp
@@ -47,7 +47,7 @@ struct AisRecord {
     // ── Match_AI (DL 지문; ais_ai.cpp 데몬 질의 결과) ──
     uint8_t  ai_status = 0;       // 0=off/데몬부재/모델없음 1=UNKNOWN 2=match
     uint32_t ai_mmsi = 0;         // 예측 MMSI (status=2)
-    uint8_t  ai_conf = 0;         // 신뢰 percent 0..100
+    uint16_t ai_conf = 0;         // 신뢰 millipercent 0..100000 (percent*1000, 소수점 3자리)
 };
 
 // wire 포맷 (framework BEWE_MK_DATA payload; station 은 MpData 봉투가 운반)
@@ -66,7 +66,7 @@ struct __attribute__((packed)) AisWireMsg {
     float    cfo_hz, fdev_std_hz, rssi_db, clk_ppm, dur_ms;
     uint8_t  spoof_flag; float cfo_z; uint32_t match_mmsi; float match_conf;
     // ── Match_AI (끝에 추가; append-only 유지) ──
-    uint8_t  ai_status; uint32_t ai_mmsi; uint8_t ai_conf;
+    uint8_t  ai_status; uint32_t ai_mmsi; uint16_t ai_conf;
 };
 
 inline void ais_msg_to_wire(const AisRecord& m, AisWireMsg& w){
